requirements.txt
pytest~=8.3.5
selenium~=4.33.0
requests~=2.32.3
python-dotenv~=1.1.0
jsonschema~=4.23.0
allure-pytest